extract RendererBase and make Renderer an interfacr
